Made some major changes to how loaded models are parsed.

+++ xkill-renderer/ViewportData.h
@@ -14,8 +14,8 @@ struct ViewportData
 	DirectX::XMFLOAT3	eyePos;
 	unsigned int		viewportTopX;
 	unsigned int		viewportTopY;
-	unsigned int		zNear;
-	unsigned int		zFar;
+	float		zNear;
+	float		zFar;
 };
 
 #endif //XKILL_RENDERER_VIEWPORTDATA_H
